Prevent data race from being triggered in commandrunner test
If you try to read from the buffers passed into Runner.Run(), you may
trigger a data race. As is the expectation of exec.Command, it is up to
the consumer of this package to be diligent about race conditions and
wait for the command to complete before reading from stdout or stderr.

[#124916101]