Implement the ServerCertDomainSAN
As a followup to the changes proposed by GOVAU, we'd like to implement a
server_cert_domain_san property to support the one from the gorouter.

We've upgraded gorouter and route-registrar to use TLS for the UAA
routes, gorouter has been failing to connect because it couldn't
validate the IP it’s connecting to against the subjectAltNames in the
certificate from Nginx on uaa/*. At the moment that cert is issued for
`uaa.service.cf.internal` which is provided by Consul service discovery
and we needed to make sure Gorouter knows about it.